Filter outdated gems to the current version of Ruby
Previously, `bundle outdated` would show the latest version of a gem,
but it didn't check whether this gem could be used in the current
version of Ruby. This resulted in some false-positives.

This commit filters out those gems that couldn't be used on the current
version of Ruby.

It's not foolproof, however, because we're only checking the
requirements of the gem and not its dependencies. This means that we may
still show versions that couldn't be installed on this version of Ruby.
For example, the `factory_bot_rails` gem doesn't specify a required Ruby
version, even though its dependency `factory_bot` specifies one.